I18n: use Language model consistently, fix admin translation save
We originally started with I18n and continued to use I18n for viewing
translations in the admin interfac even after introducing the Language
model.  Switched admin interface to rely solely on Language model.

Translations were also not getting saved when set through the admin
interface.  This was because we were modifying a hash in place which
was causing the underlying active record layer to not notice the
translation data had been changed.  resolves #29